requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-WTF==1.2.2
WTForms==3.2.1
python-dotenv==1.0.1
jiwer==3.0.4
sacrebleu==2.4.3
Pillow==11.1.0
